requirements.txt
groq==0.11.0
numpy==2.1.2
pandas==2.2.3
pdfkit==1.0.0
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.39.0
streamlit-pdf-viewer==0.0.18
urllib3==2.2.3
